requirements.txt
dspy==3.0.0b1
openai
python-dotenv
ruff